+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/CustomOp.java
@@ -82,7 +82,7 @@ public interface CustomOp {
      * Calculate the output shape for this op
      * @return
      */
-    List<int[]> calculateOutputShape();
+    List<long[]> calculateOutputShape();
 
     /**
      * Get the custom op descriptor if one is available.
